requirements.txt
# Core Data Processing
pandas>=2.0.0
numpy>=1.24.0
psutil>=5.8.0

# Data Collection APIs
praw>=7.7.0
wikipedia-api>=0.6.0
aiohttp>=3.8.0
aiofiles>=23.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# Text Processing & Cleaning
better_profanity>=0.7.0
nltk>=3.8.0

# Machine Learning
torch>=2.0.0
transformers>=4.30.0
tokenizers>=0.13.0
datasets>=2.14.0
scikit-learn>=1.3.0

# Distributed Training
deepspeed>=0.10.0
accelerate>=0.20.0

# Google Cloud Storage
google-cloud-storage>=2.9.0
google-resumable-media>=2.5.0
google-auth>=2.22.0

# Monitoring and Visualization
wandb>=0.15.0
tensorboard>=2.13.0

# Large Dataset Processing
dask>=2023.1.0
pyarrow>=12.0.0
huggingface_hub>=0.16.0
zstandard>=0.19.0

# Utilities
loguru>=0.7.0
tqdm>=4.65.0
pyyaml>=6.0.0
jsonlines>=3.1.0
memory_profiler>=0.60.0
tabulate>=0.9.0

